Add files via upload
This is the R code that did the data processing for the Flu Interactive Dashboard! I made it a quarto file in case python would be easier for any geospatial problems, but everything was simple in R.

Note: I used R to manually add the spatial attributes of the states to the CDC data. This allowed me to circumvent any credit usage by joining the spatial pieces to the database, and then add it to GIS.